backend/requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
youtube-transcript-api
